Add some text for the introduction
The whole thing needs to be reorganized - I have the feeling that I'm
missing the big picture in this paper. It would be a perfect time to do
that this week while I've got my presentations fresh in my head. Best to
use the computer and some good ol' pencil and paper.