Add scripts and update document
Scripts are now there to convert XPIF to GeoTIFF, not that we need them. Also, a
script to combine imagery from all geosynchronous satellites to a single global
image is provided. It does it in three-hourly resolution.

I also updated the thesis text by adding the validation equations. Text must
still be formulated nicely, right now it's just thoughts.